Entire pipeline working
In this version we are able to run the entire SFE pipeline, i.e., build
the graph representation, extract features from it and save features to
disk in the same protocol that the original SFE implementation used.
This implementation is working very fast, even still running on a single
core.